refactor: BoardConverter를 DB 변환 기능, Board 변환 기능으로 분리
